can't delete yourself

+++ src/main/java/com/example/oasip_back_nontonchao/services/UserService.java
@@ -1,12 +1,16 @@
 package com.example.oasip_back_nontonchao.services;
 
+import com.azure.core.annotation.Get;
 import com.example.oasip_back_nontonchao.dtos.UserGet;
 import com.example.oasip_back_nontonchao.dtos.UserUpdate;
 import com.example.oasip_back_nontonchao.entities.EventCategoryOwner;
 import com.example.oasip_back_nontonchao.entities.User;
 import com.example.oasip_back_nontonchao.repositories.EventCategoryOwnerRepository;
 import com.example.oasip_back_nontonchao.repositories.UserRepository;
+import com.example.oasip_back_nontonchao.utils.JwtTokenUtil;
 import com.example.oasip_back_nontonchao.utils.ListMapper;
+import lombok.Getter;
+import lombok.Setter;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
@@ -16,6 +20,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.Optional;
 
@@ -35,6 +40,14 @@ public class UserService {
 
     @Autowired
     private EventCategoryOwnerRepository eventCategoryOwnerRepository;
+
+    @Autowired
+    @Getter
+    @Setter
+    private HttpServletRequest httpServletRequest;
+
+    @Autowired
+    private JwtTokenUtil jwtTokenUtil;
 
     public List<UserGet> getAllUsers() {
         return listMapper.mapList(userRepository.findAll(Sort.by(Sort.Direction.ASC, "name")), UserGet.class, modelMapper);
@@ -57,14 +70,17 @@ public class UserService {
     public ResponseEntity deleteUser(Integer id) {
         Optional<User> user = userRepository.findById(id);
         if (user.get().getRole().equals("lecturer")) {
-
             if(isOnlyOne(eventCategoryOwnerRepository.getEventCategoryOwnersByUserId(id))){
                 eventCategoryOwnerRepository.deleteAssociateByUserId(id);
             }else{
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("eventCategory owner should have at least 1!");
             }
-
             }
+        String email = jwtTokenUtil.getUsernameFromToken(httpServletRequest.getHeader("Authorization").substring(7));
+        User s = userRepository.findUserByEmail(email);
+        if(s.getId() == id){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("can't delete yourself");
+        }
         userRepository.deleteById(id);
         return ResponseEntity.status(HttpStatus.OK).body("user id: " + id + " deleted!");
         }
